Got login/signup working with Async/await
