+++ readline/commands/index.js
@@ -1,0 +1,5 @@
+const login = require('./login');
+const signup = require('./signup');
+const charCreate = require('./charCreate');
+
+module.exports = { login, signup, charCreate };
