+++ readline/readline.js
@@ -10,7 +10,7 @@ const rl = readline.createInterface({
 });
 
 rl.on('line', (command) => {
-  switch (command.trim()) {
+  switch (command.trim().toUpperCase()) {
     case '1':
       console.log('CHARACTERS');
       break;
@@ -18,30 +18,18 @@ rl.on('line', (command) => {
       console.log('Entered character creator. \n');
       commands.createCharacter();
       break;
-    case 'exit':
+    case 'X':
       rl.close();
-      process.exit();
       break;
     default:
       console.log('Command not recognized');
   }
 });
 
-// Adding boolean for hiding passwords
-rl.secret = false;
-
-// modifying rl write function to be able to hide password
-rl._writeToOutput = function _writeToOutput(stringToWrite) {
-  if (rl.secret) rl.output.write('*');
-  else rl.output.write(stringToWrite);
-};
-
 // promise/async questions function
-rl.ask = (prompt, secret) => {
+rl.ask = (prompt) => {
   return new Promise((resolve, reject) => {
-    rl.secret = secret;
     rl.question(prompt, (answer) => {
-      rl.secret = false;
       resolve(answer);
     });
   });
